services/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
pydantic==2.6.4
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
